Add files via upload
In this project, we worked with the three sets of data. To start, we cleaned and organized the first two sets of data, getting them ready for analysis. We looked at basic information like the beginning, ends and unique values etc. Then we combined these two sets of data into one.
After that, we moved on to the third set of the data. Here we continued cleaning and organizing, making sure everything was in order. We also checked for any NaN values and fill them. 
I also did data profiling of all the three sets of the data.